add state tests
